requirements.txt
pynput==1.7.1
gphoto2==2.2.2
Pillow==7.1.2
pyserial==3.4
tornado==6.0.4
